4.21: bloglist expansion, step10

+++ part4/blog/controllers/blogs.js
@@ -7,10 +7,11 @@ blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user',{'username': 1,'name': 1, 'id': 1})
   response.json(blogs.map(blog => blog.toJSON()))
 })
-  
+
+
 blogsRouter.post('/', async (request, response, next) => {
   if(request.body.title && request.body.url){
-    
+
     try {
       const decodedToken = jwt.verify(request.token, process.env.SECRET)
       if (!request.token || !decodedToken.id) {
@@ -35,10 +36,29 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if(blog){
+      if(blog.user.toString() === decodedToken.id){
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      }
+      else {
+        response.status(400).json({error: 'User not authorised to delete this post'})
+      }
+    }
+
+  } catch(exception) {
+    next(exception)
+  }
 })
+
 
 blogsRouter.put('/:id', async (request, response) => {
   const result = await Blog.findByIdAndUpdate(request.params.id,{'author': request.body.author, 'url': request.body.url, 'title': request.body.title, 'likes': request.body.likes})
@@ -47,6 +67,7 @@ blogsRouter.put('/:id', async (request, response) => {
   else
     response.status(404).end()
 })
+
 
 blogsRouter.get('/:id', async (request, response) => {
   const blogs = await Blog.findById(request.params.id)
